+++ Sources/Nuke/Processing/ImageProcessors+Resize.swift
@@ -80,8 +80,8 @@ struct ImageTargetSize: Hashable {
 
     var cgSize: CGSize { CGSize(width: Double(width), height: Double(height)) }
 
-    init(maxPixelSize: Float) {
-        (width, height) = (maxPixelSize, 0)
+    init(maxPixelSize: CGFloat) {
+        (width, height) = (Float(maxPixelSize), 0)
     }
 
     /// Creates the size in pixels by scaling to the input size to the screen scale
